requirements.txt
paho-mqtt
rich
click
